rename parent route to controller route
